Add internal events to better handle some features

+++ src/main/java/ecse321/fall2014/group3/bomberman/physics/entity/mob/enemy/Balloom.java
@@ -35,7 +35,7 @@ public class Balloom extends Enemy {
     }
 
     @Override
-    public float getScore() {
+    public int getScore() {
         return 100;
     }
 }
