+++ src/main/java/ecse321/fall2014/group3/bomberman/physics/entity/mob/enemy/Pass.java
@@ -5,7 +5,6 @@ import com.flowpowered.math.vector.Vector2f;
 import ecse321.fall2014.group3.bomberman.nterface.SpriteInfo;
 import ecse321.fall2014.group3.bomberman.physics.ai.AI;
 import ecse321.fall2014.group3.bomberman.physics.ai.DumbAI;
-import ecse321.fall2014.group3.bomberman.physics.ai.SmartAI;
 
 public class Pass extends Enemy {
     private static final SpriteInfo PASS_ENEMY_SPRITE = new SpriteInfo("Sprite", 38, Vector2f.ONE);
@@ -36,7 +35,7 @@ public class Pass extends Enemy {
     }
 
     @Override
-    public float getScore() {
+    public int getScore() {
         return 4000;
     }
 }
